libs: compatible: fix warnings

+++ libs/compatible/misc.c
@@ -71,7 +71,7 @@ Balloc
 #else
     len = (sizeof(Bigint) + (x-1)*sizeof(ULong) + sizeof(double) - 1)
       /sizeof(double);
-    if (pmem_next - private_mem + len <= PRIVATE_mem) {
+    if (pmem_next - private_mem + len <= (ssize_t)PRIVATE_mem) {
       rv = (Bigint*)(void *)pmem_next;
       pmem_next += len;
       }
